Remove unnecessary casts

+++ DataInput/clsDataImportMSXml.cs
@@ -204,7 +204,7 @@ namespace MASIC.DataInput
                 return 0;
             }
 
-            return (double)ComputePrecursorInterference(
+            return ComputePrecursorInterference(
                 scanInfo.ScanNumber,
                 precursorScanNumber, parentIonMz, isolationWidth, chargeState);
         }
